nyrt: split lib into modules
